Fix 10 compiler warnings of type: sketch/menu.cpp: In function 'void _menu_edit_int3()':
sketch/menu.cpp:272:51: warning: dereferencing type-punned pointer will break strict-aliasing rules [-Wstrict-aliasing]
 #define _menu_data (*((menu_data_edit_t*)menu_data))
